Create main_tuning.py
This is largely a copy of main.py but with some changes made to allow hyperparameter tuning, which require the use of an edited version of utils.core_utils.py called utils.core_utils_tuning.py. It should be possible to integrate these in to the main versions of main.py and utils.core_utils.py in future iterations.